# api/openapi/stacks.yaml
openapi: 3.0.3
info:
  title: FloatLab Stacks API
  version: 0.1.0
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        maxLength: 255
      description: Unique key used to safely retry this mutation.
  schemas:
    StackState:
      type: string
      enum:
        - Provisioning
        - Idle
        - Starting
        - RunningPrimary
        - FailingOver
        - RunningBackup
        - Restoring
        - Stopping
        - Updating
        - RollingBack
        - Failed
      description: Current lifecycle state of the stack.
    FailoverMode:
      type: string
      enum: [manual, auto]
    Stack:
      type: object
      required: [id, name, state, primary_node, compose_file, dataset_path, failover_mode, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "plex"
        state:
          $ref: '#/components/schemas/StackState'
        primary_node:
          type: string
          description: Node ID of the primary host.
          example: "node-alpha"
        secondary_node:
          type: string
          nullable: true
          description: Node ID of the failover host.
          example: "node-beta"
        compose_file:
          type: string
          description: Raw docker-compose.yml content (includes x-fl-* extensions).
        dataset_path:
          type: string
          description: ZFS dataset path — e.g. floatlab/plex
          example: "floatlab/plex"
        failover_mode:
          $ref: '#/components/schemas/FailoverMode'
        auto_trigger_after:
          type: string
          description: Duration string, only used when failover_mode=auto.
          example: "120s"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      required: [status, message]
      properties:
        status:
          type: integer
        message:
          type: string
        code:
          type: string
paths:
...
  /stacks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get a stack by ID
      operationId: getStack
      tags: [stacks]
      responses:
        '200':
          description: Stack detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stack'
        '404':
          description: Stack not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a stack
      operationId: deleteStack
      tags: [stacks]
      description: Stops the stack and removes active metadata. The project dataset and configuration are retained unless purged.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - name: purge
          in: query
          schema:
            type: boolean
            default: false
          description: Also destroy the ZFS dataset and all snapshots.
      responses:
        '202':
          description: Delete operation accepted.
        '409':
          description: Stack is not in a deletable state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...